dbbuilder/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zemas-db-builder"
version = "0.1.0"
description = "Knowledge base builder for ZEMAS - ZEISS EUV Multi-Agent Support"
requires-python = ">=3.11"
dependencies = [
    # UI
    "PySide6>=6.7.0",

    # PDF
    "pymupdf4llm>=0.0.10",
    "PyMuPDF>=1.24.0",
    "pdfplumber>=0.11.0",

    # Excel
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",

    # Word
    "python-docx>=1.1.0",

    # OCR (default)
    "pytesseract>=0.3.10",

    # File type detection
    "python-magic>=0.4.27",

    # Embedding & VectorDB
    "httpx>=0.27.0",
    "chromadb>=0.5.0",
    "tiktoken>=0.7.0",

    # Data & Config
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",

    # NLP
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
ocr-paddle = [
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
db-builder = "db_builder.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests that take >5s (OCR, real PDF parsing)",
    "integration: marks tests requiring API keys or ChromaDB",
]

[tool.hatch.build.targets.wheel]
packages = ["src/db_builder"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
